config option `absoluteResourceURL` - use CSS/Javascript resources from the same absolute location;
this is useful for multi-module builds to avoid each submodule referencing its own style/js files, which are copied to each subdirectory. Indicate the absolute URL and relative links will be resolved based on project URL that point to the same absolute resources. Now style customization needs to be defined at the top level only and the children will reuse it.